factor out the readonly input used in Bulk Update and Bulk Delete modals
